Integrate multipart/related into multipart serializer
