+++ lib/braintreehttp/encoder.js
@@ -3,14 +3,13 @@
 let Json = require('./serializer/json').Json;
 let Text = require('./serializer/text').Text;
 let Multipart = require('./serializer/multipart').Multipart;
-let MultipartRelated = require('./serializer/multipart_related').MultipartRelated;
 let FormEncoded = require('./serializer/form_encoded').FormEncoded;
 let zlib = require('zlib');
 
 class Encoder {
 
   constructor() {
-    this._encoders = [new Json(), new Text(), new MultipartRelated(), new Multipart(), new FormEncoded()];
+    this._encoders = [new Json(), new Text(), new Multipart(), new FormEncoded()];
   }
 
   serializeRequest(request) {
