+++ lib/braintreehttp/serializer/multipart.js
@@ -9,21 +9,24 @@ class Multipart {
   }
 
   encode(request) {
+    let buffers = null;
+    let fileBuffers = [];
+    let valueBuffers = [];
     let boundary = 'boundary' + Date.now();
 
     request.headers['Content-Type'] += `; boundary=${boundary}`;
-
-    let buffers = [];
 
     for (const key of Object.keys(request.body)) {
       let val = request.body[key];
 
       if (val instanceof fs.ReadStream) {
-        buffers.push(this._filePart(key, val, boundary));
+        fileBuffers.push(this._filePart(key, val, boundary, request));
       } else {
-        buffers.push(this._formPart(key, val, boundary));
+        valueBuffers.push(this._formPart(key, val, boundary, request));
       }
     }
+
+    buffers = valueBuffers.concat(fileBuffers);
 
     buffers.push(Buffer.from(`--${boundary}--`));
     buffers.push(Buffer.from(Multipart._CRLF));
@@ -40,22 +43,38 @@ class Multipart {
     return /^multipart\/.*/;
   }
 
-  _filePart(key, readStream, boundary) {
+  _filePart(key, readStream, boundary, request) {
     return Buffer.concat([
-      Buffer.from(this._partHeader(key, path.basename(readStream.path), boundary)),
+      Buffer.from(this._partHeader(key, path.basename(readStream.path), boundary, request)),
       fs.readFileSync(readStream.path),
       Buffer.from(Multipart._CRLF)
     ]);
   }
 
-  _formPart(key, formPart, boundary) {
+  _formPart(key, formPart, boundary, request) {
+    let formPartContentType = null;
+    let formPartValue = null;
+    let contentBuffer = null;
+
+    if (formPart instanceof FormPart) {
+      formPartContentType = formPart.headers['Content-Type'];
+      formPartValue = formPart.value;
+    }
+
+    if (formPartContentType === 'application/json') {
+      contentBuffer = Buffer.from(JSON.stringify(formPartValue) + Multipart._CRLF)
+    } else {
+      contentBuffer = Buffer.from(formPart + Multipart._CRLF)
+    }
+
     return Buffer.concat([
-      Buffer.from(this._partHeader(key, null, boundary)),
-      Buffer.from(formPart + Multipart._CRLF)
+      Buffer.from(this._partHeader(key, null, boundary, request)),
+      contentBuffer
     ]);
   }
 
-  _partHeader(key, filename, boundary) {
+  _partHeader(key, filename, boundary, request) {
+    let formPart = request.body[key];
     let part = `--${boundary}`;
 
     part += Multipart._CRLF;
@@ -65,6 +84,19 @@ class Multipart {
       part += `; filename="${filename}"`;
       part += Multipart._CRLF;
       part += `Content-Type: ${this._filetype(filename)}`;
+    }
+
+    if (formPart instanceof FormPart) {
+      let partHeaders = formPart.headers;
+
+      if (partHeaders['Content-Type'] === 'application/json') {
+        part += `; filename="${key}.json"`;
+      }
+
+      for (const headerKey of Object.keys(partHeaders)) {
+        part += Multipart._CRLF;
+        part += headerKey + ': ' + partHeaders[headerKey];
+      }
     }
 
     part += `${Multipart._CRLF}${Multipart._CRLF}`;
@@ -79,6 +111,8 @@ class Multipart {
       return 'image/jpeg';
     } else if (ext === '.png') {
       return 'image/png';
+    } else if (ext === '.gif') {
+      return 'image/gif';
     } else if (ext === '.pdf') {
       return 'application/pdf';
     }
@@ -87,4 +121,19 @@ class Multipart {
   }
 }
 
-module.exports.Multipart = Multipart;
+class FormPart {
+  constructor(value, headers) {
+    this.headers = headers
+    this.value = value;
+
+    Object.keys(headers).forEach( (key) => {
+      // downcase, split into words, title case words, join into 'Header-Case'
+      this.headers[key.toLowerCase().split('-').map((word) => { return word[0].toUpperCase() + word.slice(1); }).join('-')] = headers[key];
+    });
+  }
+}
+
+module.exports = {
+  Multipart: Multipart,
+  FormPart: FormPart
+};
